+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## 0.0.98
+- CHANGED: Improve error handling
+
+## 0.0.97
+- CHANGED: Upgrade to sdui-0.1.96 to handle widget ids.
+
+## 0.0.96
+- CHANGED: Upgrade to flutter v3.0.5
+
+## 0.0.95
+- CHANGED: Disabled analytics module
+- CHANGED: Update error screen for offline mode
+
 ## 0.0.94
 - CHANGED: upgrade to sdui 0.1.94